Added function check_color_hex(color_code).
This function checks the validity of the hex code of color attributes.
A hex code is valid if it is of the form '#XXX' or '#XXXXXX'.

The 'validate_json()' function calls this function through
'validate_node_properties()' and 'validate_edge_properties()' for
node and edge color attributes repectively to ensure that the
hexadecimal codes are valid.